# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Takeovers CoS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /today/brief:
    get:
      operationId: getTodayBrief
      tags: [today]
      summary: Get today's brief
      description: Returns the current day's brief (morning or evening), generating if needed
      responses:
        "200":
          description: Today's brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Briefing"
        "500":
          description: Error generating brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /today/tasks:
    get:
      operationId: getTodayTasks
      tags: [today]
      summary: Get prioritized task list for today
      description: Returns tasks owned by the principal, sorted by priority and due date
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [captured, dispatched, acknowledged, in_progress, blocked, complete, cancelled, open, done]
      responses:
        "200":
          description: Today's task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /today/recent-captures:
    get:
      operationId: getTodayRecentCaptures
      tags: [today]
      summary: Get recent captures stream
      description: Returns the last 20 captured messages (voice + text) for quick review
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent captures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
  /webhooks/twilio/sms-inbound:
    post:
      operationId: twilioSmsInbound
      tags: [webhooks]
      summary: Twilio inbound SMS webhook
      description: |
        Validates Twilio signature, parses the inbound SMS from the principal,
        routes it through the capture pipeline, and replies via SMS.
      responses:
        "200":
          description: Webhook processed
          content:
            text/xml:
              schema:
                type: string
        "403":
          description: Invalid signature or unauthorized sender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /threads/{id}/messages:
    get:
      operationId: getThreadMessages
      tags: [threads]
      summary: Get messages in a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "404":
          description: Thread not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Briefing:
      type: object
      required: [id, date, markdown, generatedAt]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        markdown:
          type: string
        generatedAt:
          type: string
          format: date-time
        openTasksCount:
          type: ["integer", "null"]
        escalationCount:
          type: ["integer", "null"]
...
    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, done, blocked, captured, dispatched, acknowledged, complete, cancelled]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, critical, urgent, normal, null]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        dueAt:
          type: ["string", "null"]
          format: date-time
        authorityTier:
          type: ["string", "null"]
          enum: [A, B, C, null]
        tags:
          type: ["array", "null"]
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Message:
      type: object
      required: [id, threadId, role, content, createdAt]
      properties:
        id:
          type: integer
        threadId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        audioObjectPath:
          type: ["string", "null"]
        direction:
          type: ["string", "null"]
          enum: [inbound, outbound]
        senderType:
          type: ["string", "null"]
          enum: [principal, agent, team_member, external]
        contentType:
          type: ["string", "null"]
          enum: [text, voice, image, file, system]
        contentUrl:
          type: ["string", "null"]
        transcriptionConfidence:
          type: ["number", "null"]
        claudeParse:
          type: ["object", "null"]
        createdAt:
          type: string
          format: date-time
...